+++ pytket/extensions/cutensornet/structured_state/mps.py
@@ -160,9 +160,54 @@ class MPS(StructuredState):
 
         Raises:
             RuntimeError: If the ``CuTensorNetHandle`` is out of scope.
-            RuntimeError: If gate acts on more than 2 qubits or acts on non-adjacent
+            ValueError: If the command introduced is not a unitary gate.
+            ValueError: If gate acts on more than 2 qubits or acts on non-adjacent
                 qubits.
-            RuntimeError: If physical bond dimension where gate is applied is not 2.
+        """
+        try:
+            unitary = gate.op.get_unitary()
+        except:
+            raise ValueError("The command introduced is not unitary.")
+
+        # Load the gate's unitary to the GPU memory
+        unitary = unitary.astype(dtype=self._cfg._complex_t, copy=False)
+        unitary = cp.asarray(unitary, dtype=self._cfg._complex_t)
+
+        self._logger.debug(f"Applying gate {gate}.")
+        if len(gate.qubits) not in [1, 2]:
+            raise ValueError(
+                "Gates must act on only 1 or 2 qubits! "
+                + f"This is not satisfied by {gate}."
+            )
+
+        self.apply_unitary(unitary, gate.qubits)
+
+        return self
+
+    def apply_unitary(
+        self, unitary: cp.ndarray, qubits: list[Qubit]
+    ) -> StructuredState:
+        """Applies the unitary to the specified qubits of the StructuredState.
+
+        Note:
+            It is assumed that the matrix provided by the user is unitary. If this is
+            not the case, the program will still run, but its behaviour is undefined.
+
+        Args:
+            unitary: The matrix to be applied as a CuPy ndarray. It should either be
+                a 2x2 matrix if acting on one qubit or a 4x4 matrix if acting on two.
+            qubits: The qubits the unitary acts on. Only one qubit and two qubit
+                unitaries are supported.
+
+        Returns:
+            ``self``, to allow for method chaining.
+
+        Raises:
+            RuntimeError: If the ``CuTensorNetHandle`` is out of scope.
+            ValueError: If the number of qubits provided is not one or two.
+            ValueError: If the size of the matrix does not match with the number of
+                qubits provided.
+            ValueError: If qubits are non-adjacent.
         """
         if self._lib._is_destroyed:
             raise RuntimeError(
@@ -170,38 +215,37 @@ class MPS(StructuredState):
                 "See the documentation of update_libhandle and CuTensorNetHandle.",
             )
 
-        positions = [self.qubit_position[q] for q in gate.qubits]
-        if any(self.get_physical_dimension(pos) != 2 for pos in positions):
-            raise RuntimeError(
-                "Gates can only be applied to tensors with physical"
-                + " bond dimension of 2."
-            )
-        self._logger.debug(f"Applying gate {gate}")
+        self._logger.debug(f"Applying unitary {unitary} on {qubits}.")
 
-        if len(positions) == 1:
-            self._apply_1q_gate(positions[0], gate.op)
+        if len(qubits) == 1:
+            if unitary.shape != (2, 2):
+                raise ValueError(
+                    "The unitary introduced acts on one qubit but it is not 2x2."
+                )
+
+            self._apply_1q_unitary(unitary, qubits[0])
             # NOTE: if the tensor was in canonical form, it remains being so,
             #   since it is guaranteed that the gate is unitary.
 
-        elif len(positions) == 2:
+        elif len(qubits) == 2:
+            if unitary.shape != (4, 4):
+                raise ValueError(
+                    "The unitary introduced acts on two qubits but it is not 4x4."
+                )
+
+            positions = [self.qubit_position[q] for q in qubits]
             dist = positions[1] - positions[0]
             # We explicitly allow both dist==1 or dist==-1 so that non-symmetric
             # gates such as CX can use the same Op for the two ways it can be in.
             if dist not in [1, -1]:
-                raise RuntimeError(
-                    "Gates must be applied to adjacent qubits! "
-                    + f"This is not satisfied by {gate}."
-                )
-            self._apply_2q_gate((positions[0], positions[1]), gate.op)
+                raise ValueError("Gates must be applied to adjacent qubits!")
+            self._apply_2q_unitary(unitary, qubits[0], qubits[1])
             # The tensors will in general no longer be in canonical form.
             self.canonical_form[positions[0]] = None
             self.canonical_form[positions[1]] = None
 
         else:
-            raise RuntimeError(
-                "Gates must act on only 1 or 2 qubits! "
-                + f"This is not satisfied by {gate}."
-            )
+            raise ValueError("Gates must act on only 1 or 2 qubits!")
 
         return self
 
@@ -919,13 +963,13 @@ class MPS(StructuredState):
         """
         return len(self.tensors)
 
-    def _apply_1q_gate(self, position: int, gate: Op) -> MPS:
+    def _apply_1q_unitary(self, unitary: cp.ndarray, qubit: Qubit) -> MPS:
         raise NotImplementedError(
             "MPS is a base class with no contraction algorithm implemented."
             + " You must use a subclass of MPS, such as MPSxGate or MPSxMPO."
         )
 
-    def _apply_2q_gate(self, positions: tuple[int, int], gate: Op) -> MPS:
+    def _apply_2q_unitary(self, unitary: cp.ndarray, q0: Qubit, q1: Qubit) -> MPS:
         raise NotImplementedError(
             "MPS is a base class with no contraction algorithm implemented."
             + " You must use a subclass of MPS, such as MPSxGate or MPSxMPO."